chore: put difficulty param mixins into new package
